refactor: MVC 리팩터링
