+++ src/main/java/racing/controller/CarController.java
@@ -9,43 +9,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class CarController {
-    private final CarView carView;
+    private final RacingGame racingGame;
 
-    public CarController() {
-        carView = new CarView();
+    public CarController(RacingGame racingGame) {
+        this.racingGame = racingGame;
     }
 
-    public void run() {
-        String inputCarNames = carView.getCarNames();
-        String[] carNames = inputCarNames.split(",");
-        RacingGame racingGame = new RacingGame(createCarList(carNames));
-
-        int turn = carView.getTryCount();
-
+    public void run(int turn) {
         for (int i = 0; i < turn; i++) {
             racingGame.proceed();
         }
-
-        carView.printResultDescription();
-        for (int time = 0; time < turn; time++) {
-            carView.printCarList(racingGame, time);
-        }
-
-        List<Car> cars = racingGame.getCarList();
-        int maxPosition = racingGame.getMaxPosition(cars);
-        List<Car> winners = racingGame.getWinner(cars, maxPosition);
-
-        carView.printWinner(winners);
-    }
-
-    private List<Car> createCarList(String[] carNames) {
-        List<Car> carList = new ArrayList<>();
-        RandomNumberGenerator randomNumberGenerator = new RandomNumberGenerator();
-        for (String carName : carNames) {
-            Car car = new Car(carName, randomNumberGenerator);
-            carList.add(car);
-        }
-
-        return carList;
     }
 }
